Preparing for DSM 7 compatibility

+++ src/make.sh
@@ -3,13 +3,13 @@
 
 . `dirname "$0"`/INFO
 SPK_FILE="${package}_${version}.spk"
+SPK_FILE_7="${package}_${version}_DSM7.spk"
 
 
 pack_perl()
 {
 	# remove comments
 	sed -r -i '/^\s*#[^!].*?$|^\s*#$/d' "$1"
-
 }
 
 pack_javascript()
@@ -52,33 +52,26 @@ pack_php()
 
 make()
 {
-	echo "Make ${SPK_FILE}..."
+	echo "Make ${package}_${version}..."
 	if [ -f "$SPK_FILE" ]; then
 		echo "$SPK_FILE" already exists
 		exit 1
 	fi
+	if [ -f "$SPK_FILE_7" ]; then
+		echo "$SPK_FILE_7" already exists
+		exit 1
+	fi
 	
-	echo "== Copy sources"
+	echo "== Create package.tgz"
 	rm -rf _tmp
 	mkdir _tmp
-	mkdir _tmp/package
 	mkdir _tmp/spk
+	mkdir _tmp/package
 	
 	cp -af target/* _tmp/package
-	cp -af INFO _tmp/spk
-	cp -af scripts _tmp/spk
-	cp -af WIZARD_UIFILES _tmp/spk
-	if [ -d conf ]; then
-		cp -af conf _tmp/spk
-		mkdir _tmp/package/conf_pkgcenter
-		cp -af conf/* _tmp/package/conf_pkgcenter/
-	fi
 	
 	find _tmp -name '*.bak' -delete
 
-	cp -af target/ui/images/icon_72.png _tmp/spk/PACKAGE_ICON.PNG
-	
-	
 	# check perl files
 #	for srcfile in `grep -ril '^#!.*perl' ./_tmp/package/`
 	for srcfile in `find ./_tmp/package/ -type f | grep -Ev 'lib/Image/|lib/File/' | grep -E '\.(pl|pm|cgi)'`
@@ -135,20 +128,52 @@ make()
 		echo "    ${phpfile##*/} - OK"
 	done
 
-	echo "== Compress to package.tgz"
 	cd _tmp/package/
-	tar czf ../spk/package.tgz --owner='' --group='' --mode=0755 *
+	tar czf ../spk/package.tgz --owner=root --group=root *
 	cd ../..
 
+	echo "== Create $SPK_FILE"
+	
+	cp -af INFO _tmp/spk
+	cp -af scripts _tmp/spk
+	cp -af WIZARD_UIFILES _tmp/spk
+	if [ -d conf ]; then
+		cp -af conf _tmp/spk
+#		mkdir _tmp/package/conf_pkgcenter
+#		cp -af conf/* _tmp/package/conf_pkgcenter/
+	fi
+	
+	cp -af target/ui/images/icon_72.png _tmp/spk/PACKAGE_ICON.PNG
+	cp -af target/ui/images/icon_256.png _tmp/spk/PACKAGE_ICON_256.PNG
 	#echo "extractsize=\"`du -sb ./_tmp/package | cut -f 1`\"" >> _tmp/spk/INFO
 	#echo "checksum=\"` md5sum ./_tmp/spk/package.tgz | cut -f 1 -d ' '`\"" >> _tmp/spk/INFO
 	#sed -i '/^create_time\s*=/d' _tmp/spk/INFO
 	echo "create_time=\"`date +%Y%m%d-%H:%M:%S`\"" >> _tmp/spk/INFO
-	
-	echo "== Compress to $SPK_FILE"
+
 	cd _tmp/spk/
-	tar cf ../../"$SPK_FILE" --owner='' --group='' --mode=0755 *
+	tar cf ../../"$SPK_FILE" --owner=root --group=root *
 	cd ../..
+	
+	if [ -e INFO.DSM7 ]; then
+		echo "== Create $SPK_FILE_7"
+		cp -af INFO.DSM7 _tmp/spk/INFO
+		#echo "extractsize=\"`du -sb ./_tmp/package | cut -f 1`\"" >> _tmp/spk/INFO
+		#echo "checksum=\"` md5sum ./_tmp/spk/package.tgz | cut -f 1 -d ' '`\"" >> _tmp/spk/INFO
+		#sed -i '/^create_time\s*=/d' _tmp/spk/INFO
+		echo "create_time=\"`date +%Y%m%d-%H:%M:%S`\"" >> _tmp/spk/INFO
+
+		if [ -d conf.DSM7 ]; then
+			rm -rf _tmp/spk/conf
+			cp -af conf.DSM7 _tmp/spk/conf
+#			mkdir _tmp/package/conf_pkgcenter
+#			cp -af conf/* _tmp/package/conf_pkgcenter/
+		fi
+		cd _tmp/spk/
+		tar cf ../../"$SPK_FILE_7" --owner=root --group=root * #--mode=0755 *
+
+		cd ../..
+	fi
+
 }
 
 rm_rf()
@@ -162,6 +187,7 @@ rm_rf()
 make_clean()
 {
 	rm_rf "$SPK_FILE"
+	rm_rf "$SPK_FILE_7"
 	rm_rf "_tmp"
 }
